aoc-lib/src/tests/year2022/day11_test.rs
use crate::solution::*;
use crate::solutions::year2022::Day11;
use crate::tests::util::*;

#[test]
fn example_input() {
    let (mut day, ctx) = setup::<Day11>(
        r#"
Monkey 0:
  Starting items: 79, 98
  Operation: new = old * 19
  Test: divisible by 23
    If true: throw to monkey 2
    If false: throw to monkey 3

Monkey 1:
  Starting items: 54, 65, 75, 74
  Operation: new = old + 6
  Test: divisible by 19
    If true: throw to monkey 2
    If false: throw to monkey 0

Monkey 2:
  Starting items: 79, 60, 97
  Operation: new = old * old
  Test: divisible by 13
    If true: throw to monkey 1
    If false: throw to monkey 3

Monkey 3:
  Starting items: 74
  Operation: new = old + 3
  Test: divisible by 17
    If true: throw to monkey 0
    If false: throw to monkey 1"#,
    );
    assert_result(day.part1(&ctx), "10605", "solve part 1");
    assert_result(day.part2(&ctx), "2713310158", "solve part 2");
}

#[test]
fn puzzle_input() {
    let (mut day, ctx) = setup_from_file::<Day11>();
    assert_result(day.part1(&ctx), "64032", "solve part 1");
    assert_result(day.part2(&ctx), "12729522272", "solve part 2");
}
